Added support for strictly-typed states and events.
