+++ spring-statemachine-uml/src/main/java/org/springframework/statemachine/uml/support/GenericTypeConverter.java
@@ -1,0 +1,30 @@
+/*
+ * Copyright 2016-2021 the original author or authors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.springframework.statemachine.uml.support;
+
+import org.springframework.statemachine.uml.GenericUmlStateMachineModelFactory;
+import org.springframework.statemachine.uml.UmlStateMachineModelFactory;
+
+/**
+ * Required when extending from or using {@link GenericUmlStateMachineModelFactory}
+ * rather than {@link UmlStateMachineModelFactory}. Implementations will facilitate
+ * conversion between native UML representations of state and event i.e. {@link String}
+ * and your custom state/event types or enums.
+ */
+public interface GenericTypeConverter<E> {
+    E convert(String from);
+    String revert(E from);
+}
